Package.swiftの配置を変更
